==> 01_DataSampling.ipynb <==
#comment begin
#using keywords to filter covid-19 tweets
#using EU_countries list to filter tweets originating from EU countries
#using null value of reply id to filter original tweets
#Filter original, covid, eu tweets from the dataset
#comment end



import pandas as pd

keywords = ['covid-19', 'covid', 'pandemic', 'epidemic', 'wuhan', 'incidence', 'socialdistancing', 'social distancing',
                'incubation', 'isolation', 'mers', 'sars', 'social distance', 'panic', 'panicbuy', 'panic buy', 'panic buying',
                'quarantine', 'respirator', 'virus', 'corona', 'coronavirus', 'lockdown', 'quarantined', 'corona crisis', 'stayhome',
           'stay home', 'wash your hands', 'washyourhands','staysafe','pandemia','health','stayhealthy','quarantinelife',
            'coronamemes']

EU_countries = ['Austria','Belgium','Bulgaria','Croatia','Republic of Cyprus','Czech Republic','Denmark','Estonia','Finland',
                'France','Germany','Greece','Hungary','Ireland','Italy','Latvia','Lithuania','Luxembourg','Malta','Netherlands',
                'Poland','Portugal','Romania','Slovakia','Slovenia','Spain','Sweden']

c = 0

for df in pd.read_json("/data/twitter/ss2020/all/data7.json",chunksize=100000,lines=True):
    inp = pd.DataFrame(df, columns=['created_at','id','in_reply_to_status_id','place','reply_count','text'])
    #filter covid19 tweets
    pat1 = '|'.join(keywords)
    covid_tweet = inp.text.str.contains(pat1,case=False,regex=True)
    
    #filter tweets from list of EU countries
    EU_tweet = inp['place'].str.get('country').isin(EU_countries)
    
    #filtering with only the main tweets
    filterOriginalTweets = pd.isnull(inp['in_reply_to_status_id'])
    
    #combining the tweets that satisfy all threw conditions
    # appending data for each iteration of for loop
    if c == 0:
        covid_EU_data = inp[covid_tweet & filterOriginalTweets & EU_tweet]
        c += 1
    else:
        covid_EU_data = covid_EU_data.append(inp[covid_tweet & filterOriginalTweets & EU_tweet], ignore_index=True)

#write filtered tweets into a separate file        
covid_EU_data.to_json(r'/data/kash1240/covid_eu7.json',orient='records',lines=True)
