==> 03_DiffusionAnalysis.ipynb <==
#get reply count
import pandas as pd

c=0

for dataset in pd.read_json('/data/twitter/ss2020/all/data12.json',chunksize=100000,lines=True):
    if c == 0:
        covid_replies = dataset[['in_reply_to_status_id','created_at']].dropna()
    else:
        covid_replies = covid_replies.append(dataset[['in_reply_to_status_id','created_at']].dropna(), ignore_index=True)

covid_replies['in_reply_to_status_id'] = covid_replies['in_reply_to_status_id'].astype(int)

counts = covid_replies['in_reply_to_status_id'].astype(int).value_counts()

covid_replies['Reply_Count'] = covid_replies['in_reply_to_status_id'].map(counts).fillna(0).astype(int)

covid_replies = pd.read_json("/data/kash1240/covid_replies.json",lines=True)

covid_replies.sort_values(by='Reply_Count', ascending=False)

covid_replies = covid_replies.drop_duplicates(subset = ['Reply_Count'])

covid_replies.rename(columns={"created_at": "Reply_Time"})

#read original tweet
import pandas as pd

orig_tweet = pd.read_json('/data/kash1240/covideu_final.json',lines=True)

#mapping reply count and reply time to the original message
orig_tweet = (orig_tweet.merge(covid_replies, left_on='id', right_on='in_reply_to_status_id')
          .reindex(columns=['id', 'created_at', 'place', 'polarity','polarity_score', 'Reply_Count', 'Reply_Time']))
