==> 04_Result_Analysis.ipynb <==
df.to_json(r'/data/kash1240/covideu_final.json',orient='records',lines=True)

df1 = pd.read_json("/data/kash1240/covideu_final.json",lines=True)

#separating positive,negative and neutral tweets
column_names=['created_at','id','in_reply_to_status_id','place','polarity','polarity_score','reply_count','text']
dfo = pd.DataFrame(columns=column_names)
dfp = pd.DataFrame(columns=column_names)
dfn = pd.DataFrame(columns=column_names)

dfo = df1.loc[df1['polarity'] == 'neutral']
dfp = df1.loc[df1['polarity'] == 'positive']
dfn = df1.loc[df1['polarity'] == 'negative']

df1.shape

#number of positive tweets
dfp.shape

#number of negative tweets
dfn.shape

#number of neutral tweets
dfo.shape

#positive mean value
dfp['polarity_score'].mean()

#negative mean value
dfn['polarity_score'].mean()

import pandas as pd
pd.set_option('display.max_colwidth',-1)

import pandas as pd
df1 = pd.read_json("/data/kash1240/covideu_final.json",lines=True)

df1[['id','text','polarity_score','polarity']].head(10)
